requirements.txt
# =========================
# Web Framework
# =========================
flask==3.0.0
flask-cors==4.0.0

# =========================
# Core AI / LangChain Stack
# =========================
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.30
langchain-groq>=0.1.0
langgraph>=0.0.40
langchain-openai

# =========================
# LLM / Embeddings
# =========================
sentence-transformers>=2.3.1
numpy>=1.24.0

# =========================
# LlamaIndex (Optional)
# =========================
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-supabase>=0.1.0

# =========================
# Document Processing
# =========================
PyPDF2>=3.0.1
pypdf>=3.17.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pandas>=2.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

# =========================
# Database / Memory
# =========================
supabase>=2.3.0
vecs>=0.4.0
psycopg2-binary>=2.9.9
mem0ai>=0.1.0

# =========================
# Utilities
# =========================
pydantic>=2.7.3,<3.0.0
requests>=2.31.0
python-dotenv>=1.0.0

gunicorn>=21.2.0